update testcase. make abstract base class for python test
