+++ repl/src/main/scala/com/cloudera/livy/repl/PythonInterpreter.scala
@@ -38,7 +38,7 @@ import com.cloudera.livy.sessions._
 
 // scalastyle:off println
 object PythonInterpreter extends Logging {
-  def apply(conf: SparkConf, kind: Kind = PySpark()): Interpreter = {
+  def apply(conf: SparkConf, kind: Kind): Interpreter = {
     val pythonExec = kind match {
         case PySpark3() => sys.env.getOrElse("PYSPARK3_DRIVER_PYTHON", "python3")
         case PySpark() => sys.env.getOrElse("PYSPARK_DRIVER_PYTHON", "python")
