+++ repl/src/main/scala/com/cloudera/livy/repl/ReplDriver.scala
@@ -46,7 +46,7 @@ class ReplDriver(conf: SparkConf, livyConf: RSCConf)
 
   override protected def initializeContext(): JavaSparkContext = {
     val interpreter = Kind(livyConf.get(RSCConf.Entry.SESSION_KIND)) match {
-      case PySpark() => PythonInterpreter(conf)
+      case PySpark() => PythonInterpreter(conf, PySpark())
       case PySpark3() => PythonInterpreter(conf, PySpark3())
       case Spark() => new SparkInterpreter(conf)
       case SparkR() => SparkRInterpreter(conf)
